Minor cleanups
Clean up and split minidump code into separate files.
